playbook: Initial revision
Initial checkin of the playbook code. Playbooks are useful for
gradually automating (or semi-automating) repetitive and mundane
tasks.

The idea is similar to a checklist: have a sequence of tasks that must
be accomplished in order to complete the playbook. This is usually
implemented by starting out with a series of WaitToProceed tasks that
print some task, then wait for the user to indicate it is completed
before proceeding. As these manually tasks are gradually automated,
the WaitToProceed tasks can be replaced with code that automatically
completes the task.

The WaitToProceed task is the only "generic" task implemented so far
(more coming in the future), but tasks can still be written manually
for custom behavior.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>