AcceptUserInput: Initial revision
It is fairly common for a playbook to require the user to input some
information (e.g. path to a downloaded file, email address for data
processing, etc). Add a new abstract task type that simply asks for
the information then handles the input in the child class.

Re-implement the WaitToProceed task on top of this. This only saves
one repeated line; the main benefit here is reduced testing surface
and a proof-of-concept of AcceptUserInput children classes.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>