Playbook: Pretty-print docstrings
When printing the docstring before running a task, dedent and wrap the
docstring for a prettier format. This should nicely dedent docstrings
that are aligned with the column of the initial """ as well as
properly wrapping to a fixed column (instead of wrapping based on the
source file).

Note that this code only works if there is not end-of-line whitespace
(otherwise the split() will fail). Linting should ensure this.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>