toplevel: Add readline history file for input()
It is fairly useful to use a global history file that can be shared
through all the AcceptUserInput tasks (and anything using input()
really). On import, load the history file if it exists and register a
function to save it on exit. This should save the history even on
exceptions propagating to the toplevel, which should be useful for
replaying playbooks.

Note that it doesn't seem great to do this at the module level, but
I'm not sure there's a better choice. Loading the history file in the
Playbook class would happen too many times (since Playbook might be
constructed many times).

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>