WaitToProceed: Add tab completion
Add readline tab completion for completing all possible inputs (either
"y" or "n").

The AcceptUserInput docstring is updated to describe how the phases
should be used to set up and remove tab completion.

Tab completion is enabled globally on readline setup; this will be
useful in a number of tasks and there's no better place to put it to
run only once.

This is really just a proof-of-concept and example of using the
prepare() and cleanup() phases since it isn't very valuable to
tab-complete a single character.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>