SerialPlaybook: Remove docstring
Printing a docstring at the start of a serial playbook is rarely
useful, since the docstrings for the component playbooks are used
instead. Remove the docstring from the base class as well as the doc
argument for setting the docstring.

If I find this useful later, it may be added back later as an optional
argument.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>