AcceptPathInput: Initial revision
For a lot of playbooks, it is useful to accept a path from the user
for some additional processing. Create the AcceptPathInput task for
handling this.

The AcceptUserInput task is used to implement this, along with some
custom path completion using readline. Once the input is read from the
user, it is passed to the accept_path() method, which should be
implemented in a child class. If the path does not exist, an exception
will be thrown.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>